Changed casting of TICK env value to float instead of integer

+++ lib/resque_scheduler/tasks.rb
@@ -20,7 +20,7 @@ namespace :resque do
 
     Resque::Scheduler.dynamic = true if ENV['DYNAMIC_SCHEDULE']
     Resque::Scheduler.verbose = true if ENV['VERBOSE']
-    Resque::Scheduler.poll_sleep_amount = ENV['TICK'].to_i if ENV['TICK']
+    Resque::Scheduler.poll_sleep_amount = ENV['TICK'].to_f if ENV['TICK']
     Resque::Scheduler.run
   end
